# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "بذور الأمل Platform API",
    "version": "0.1.0",
    "description": "Generated from the implementation. The SRS TD-3 route registry is the canonical contract (§3.1); this document must conform to it and is never hand-edited."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin API path (§3.1)"
    }
  ],
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message_key",
              "message",
              "details",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "The canonical application error code (TD-3.8). Extensible only by SRS revision.",
                "enum": [
                  "VALIDATION_FAILED",
                  "AUTH_REQUIRED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "STATE_CONFLICT",
                  "VERSION_CONFLICT",
                  "DUPLICATE",
                  "CONSENT_REQUIRED",
                  "CAPACITY_FULL",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL"
                ]
              },
              "message_key": {
                "type": "string",
                "description": "i18n key; the client renders this."
              },
              "message": {
                "type": "string",
                "description": "Localised fallback text (Arabic at launch, §6)."
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "Structured context for codes that carry it — e.g. `reason: ROOM_TIME_OVERLAP` with `conflicting_group_id`, `capacity` on CAPACITY_FULL, or `conflicting_month` on a Hijri ordering refusal. Empty object when there is nothing to add."
              },
              "request_id": {
                "type": "string",
                "description": "Correlates the response with the server log line."
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/admin/hijri-calendar/{year}/{month}": {
      "put": {
        "summary": "Record the official start of one Hijri month",
        "description": "**Records the Ministry of Habous's official announcement** for this month — the Gregorian date on which it announced the month began. **The Super Admin is not deciding this date** (Revision 32); the Ministry decides it by sighting, and this transcribes it. **TD-15 optimistic locking**: `version` is required when correcting an existing month and a stale one is `409` — two Super Admins correcting the same month must not clobber each other. **A correction returns the month to `draft`**, so a change to live data must be reviewed and republished deliberately. **TD-9 ordering** is enforced across the year boundary in both directions: month *n+1* must start after month *n* and no two months may share a start date, because an out-of-order pair makes date resolution ambiguous — and resolution is what every Hijri label in the platform depends on. Dates are `YYYY-MM-DD`, local calendar dates (TD-11), never instants.",
        "responses": {
          "200": {
            "description": "The recorded month with its new version."
          },
          "400": {
            "description": "Error envelope (TD-3.8). VALIDATION_FAILED — malformed date, month outside 1–12, a missing version on an existing month, or `reason: 'MONTH_ORDER'` with the conflicting month in `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Error envelope (TD-3.8).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope (TD-3.8).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope (TD-3.8). VERSION_CONFLICT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Error envelope (TD-3.8). INTERNAL — unexpected server fault; details are never leaked to the client.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Withdraw a recorded Hijri month",
        "description": "**Revision 59.5.** `HijriMonthStart` was the only entity a Super Admin can create through the platform with **no deletion at all** — the column existed, every read filtered on it, and nothing could ever set it, so a month entered by mistake was permanent. Soft delete with a `Trash` snapshot, like every other deletion (TD-5, BR-15). **Super Admin only**, asserted in the service. **TD-15: `version` is REQUIRED**, as a query parameter rather than a body — withdrawing a month somebody else has just corrected is the identical hazard a correction guards against, and a `DELETE` with a body is refused by proxies. **Only the LAST recorded month may be withdrawn** (`LATER_MONTH_RECORDED`, naming the month that blocks it): the months are a contiguous sequence and §5.7's conversion walks it, so a hole would surface to a reader as *missing Ministry data* rather than as the deletion that caused it. **Publication is not a second obstacle** — §5.7 already treats a correction as returning a month to `draft`, so a withdrawal is the stronger form of something the specification permits; what is refused is the hole, visible or not.",
        "responses": {
          "204": {
            "description": "Withdrawn."
          },
          "400": {
            "description": "Error envelope (TD-3.8). VALIDATION_FAILED — an invalid year/month, or a missing version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Error envelope (TD-3.8).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope (TD-3.8). FORBIDDEN — Super Admin only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope (TD-3.8). NOT_FOUND when no such month is recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope (TD-3.8). STATE_CONFLICT with LATER_MONTH_RECORDED, or VERSION_CONFLICT (TD-15).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Error envelope (TD-3.8). INTERNAL — unexpected server fault; details are never leaked to the client.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}